pyproject.toml
[project]
name = "myworkflow-workspace"
version = "0.1.0"
description = "Workspace-level environment synced from old_src"
readme = "old_src/README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "langchain-core>=1.2.25",
    "langchain-openai>=1.0.1",
    "langchain-google-genai>=4.2.1",
    "loguru>=0.7.3",
    "marko>=2.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "simpleeval>=1.0.7",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "e2e: real LLM end-to-end tests",
    "slow: long-running tests",
]
